# backend/openapi.yaml
openapi: 3.0.0
info:
  title: Connectify Nigeria Provider API
  version: 1.0.0
  description: API specification for Connectify Nigeria provider endpoints
servers:
  - url: http://localhost:8000
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Job:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [pending, accepted, in_progress, completed, cancelled]
        customer:
          $ref: '#/components/schemas/Customer'
        amount:
          type: number
          format: float
        created_at:
          type: string
          format: date-time
...
paths:
...
  /api/providers/me/jobs:
    get:
      summary: Get provider jobs
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, accepted, in_progress, completed, cancelled]
      responses:
        '200':
          description: Provider jobs retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Job'
  /api/providers/me/jobs/{job_id}/status:
    put:
      summary: Update job status
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: job_id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [accepted, in_progress, completed, cancelled]
      responses:
        '200':
          description: Job status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
...